% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASICSimulationModel.R
\name{BASICSimulationModel}
\alias{BASICSimulationModel}
\title{Create an instance of \emph{BASICSimulationModel}}
\usage{
BASICSimulationModel()
}
\value{
an instance of \emph{BASICSimulationModel} class
}
\description{
Create an instance of \emph{BASICSimulationModel}
}
