% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{CSVFileParser}
\alias{CSVFileParser}
\title{Create an instance of \emph{CSVFileParser}}
\usage{
CSVFileParser(pathname)
}
\arguments{
\item{pathname}{\emph{character} containing \code{pathname}}
}
\value{
an instance of \emph{CSVFileParser} class
}
\description{
Create an instance of \emph{CSVFileParser}
}
