% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{CSVResultsExporter}
\alias{CSVResultsExporter}
\title{Create an instance of \emph{CSVResultsExporter}}
\usage{
CSVResultsExporter(pathname)
}
\arguments{
\item{pathname}{a character containing pathname}
}
\value{
an instance of \emph{CSVResultsExporter} class
}
\description{
Create an instance of \emph{CSVResultsExporter}
}
