% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rJava_utils.R
\name{.rvec2jarraylist}
\alias{.rvec2jarraylist}
\title{Add elements of an R vector to new Java ArrayList}
\usage{
.rvec2jarraylist(x, class)
}
\arguments{
\item{x}{an R vector}

\item{class}{fully qualified class name in JNI notation (e.g. "java/lang/String")}
}
\value{

}
\description{
Add elements of an R vector to new Java ArrayList
}
\examples{

.rvec2jarraylist(as.double(c(1.234, 2.345)), "java/lang/Double")

}
