% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jNSMR-package.R
\docType{package}
\name{jNSMR-package}
\alias{jNSMR}
\alias{jNSMR-package}
\title{jNSMR: Wrapper for Java Newhall Simulation Model (jNSM) "A Traditional Soil Climate Simulation Model"}
\description{
Provides methods to create input, read output, and run the routines from the legacy Java Newhall Simulation Model (jNSM). Currently this package uses version 1.6.1 of the jNSM available for download here: <https://www.nrcs.usda.gov/wps/portal/nrcs/detail/?cid=nrcs142p2_053559>. The system requirements of the extraction and installation tools (Windows .EXE archive) at the above link may not be met on your system but the core Java class files are stored in a platform-independent format (in newhall-1.6.1.jar) which is a core dependency in this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ncss-tech.github.io/jNSMR}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
