% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallData.R
\name{newhall_cmip6_cache}
\alias{newhall_cmip6_cache}
\alias{newhall_cmip6_rast}
\alias{newhall_cmip6_subset}
\title{Load CMIP6 Downscaled Future Climate Projections}
\usage{
newhall_cmip6_cache(
  model,
  ssp,
  time,
  resolution = "10m",
  version = "2.1",
  overwrite = FALSE,
  CMIP6_PATH = file.path(newhall_data_dir("cache"), "CMIP6")
)

newhall_cmip6_rast(
  model,
  ssp,
  time,
  resolution = "10m",
  version = "2.1",
  CMIP6_PATH = file.path(newhall_data_dir("cache"), "CMIP6"),
  tiffile = list.files(file.path(CMIP6_PATH, paste0("wc", version, "_", resolution)),
    pattern = "\\\\.tif$", recursive = TRUE)
)

newhall_cmip6_subset(
  x,
  model,
  ssp,
  time,
  resolution = "10m",
  template = "EPSG:4326",
  CMIP6_PATH = file.path(newhall_data_dir("cache"), "CMIP6")
)
}
\arguments{
\item{model}{\emph{character}. Climate model abbrevation. One of "ACCESS-CM2", "ACCESS-ESM1-5", "AWI-CM-1-1-MR", "BCC-CSM2-MR", "CanESM5", "CanESM5-CanOE", "CMCC-ESM2", "CNRM-CM6-1", "CNRM-CM6-1-HR", "CNRM-ESM2-1", "EC-Earth3-Veg", "EC-Earth3-Veg-LR", "FIO-ESM-2-0", "GFDL-ESM4", "GISS-E2-1-G", "GISS-E2-1-H", "HadGEM3-GC31-LL", "INM-CM4-8", "INM-CM5-0", "IPSL-CM6A-LR", "MIROC-ES2L", "MIROC6", "MPI-ESM1-2-HR", "MPI-ESM1-2-LR", "MRI-ESM2-0", "UKESM1-0-LL"}

\item{ssp}{\emph{character}. A valid Shared Socio-economic Pathway code: "126", "245", "370" or "585"}

\item{time}{\emph{character}. A valid time period. One of "2021-2040", "2041-2060", or "2061-2080"}

\item{resolution}{character. Either \code{"10m"}, \code{"5m"}, \code{"2.5m"} or \code{"30s"}. In minutes (or seconds) of degrees (\code{"EPSG:4326"}).}

\item{version}{character. Version number. Default: \code{"2.1"}. See \code{geodata::worldclim_global()} for details.}

\item{overwrite}{Force download of new cache files? Default: \code{FALSE}.}

\item{CMIP6_PATH}{Default: \code{file.path(newhall_data_dir("cache"), "CMIP6")}}

\item{tiffile}{Optional: custom vector of paths to files to use to build raster. Defaults to all .TIF files in the specified cache directory and resolution.}

\item{x}{A \emph{SpatVector}, \emph{SpatRaster}, \emph{SpatExtent}, or any other object type suitable to use with \code{terra::crop()}.}

\item{template}{Template \emph{SpatRaster} or target CRS specification for re-projection. Default: \verb{"EPSG:4326}}
}
\value{
character. Vector of file paths (to CMIP6 .TIF files).
}
\description{
\code{newhall_cmip6_cache()}: Uses the \href{https://cran.r-project.org/package=geodata}{geodata} package to download and cache data at the specified resolution.

\code{newhall_cmip6_rast()}: Create a \emph{SpatRaster} object. This object contains temperature and precipitation data for the specified data set, at the specified resolution, using the standard jNSM column naming scheme.

\verb{newhall_cmip6_subset():} Used to create a subset of the CMIP6 data corresponding to the extent of an input spatial object \code{x}.
}
\references{
\itemize{
\item Eyring, V., Bony, S., Meehl, G. A., Senior, C. A., Stevens, B., Stouffer, R. J., and Taylor, K. E.: Overview of the Coupled Model Intercomparison Project Phase 6 (CMIP6) experimental design and organization, Geosci. Model Dev., 9, 1937-1958, doi:10.5194/gmd-9-1937-2016, 2016.
\item Detailed and up-to-date description of the CMIP6 experiments protocol: https://search.es-doc.org/?project=cmip6&
}
}
\seealso{
\code{\link[=newhall_data_dir]{newhall_data_dir()}}
}
