% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{newhall_CSV_export}
\alias{newhall_CSV_export}
\title{Export Newhall Results, Data and Metadata to CSV file}
\usage{
newhall_CSV_export(pathname, dataset, results)
}
\arguments{
\item{pathname}{output CSV file path}

\item{dataset}{\emph{NewhallDataset} \code{jobjRef}}

\item{results}{\emph{NewhallResults} \code{jobjRef}}
}
\value{
an CSV file written to specified path
}
\description{
Export Newhall Results, Data and Metadata to CSV file
}
