% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{newhall_CSVResultsExporter}
\alias{newhall_CSVResultsExporter}
\title{Export Newhall Results, Data and Metadata to CSV file with \emph{CSVResultsExporter}}
\usage{
newhall_CSVResultsExporter(results, pathname)
}
\arguments{
\item{results}{\emph{NewhallResults} \code{jobjRef}}

\item{pathname}{output CSV file path; default: \code{NULL}}
}
\value{
a CSV file written to specified path
}
\description{
Export Newhall Results, Data and Metadata to CSV file with \emph{CSVResultsExporter}
}
