% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallData.R
\name{newhall_data_dir}
\alias{newhall_data_dir}
\title{Newhall Data Directory}
\usage{
newhall_data_dir(which = c("data", "config", "cache"))
}
\arguments{
\item{which}{One of: \code{"data"}, \code{"config"}, or \code{"cache"}}
}
\value{
character. Directory path.
}
\description{
Returns a platform-specific user-level directory where data, configuration and cache files may be stored.
}
