% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallData.R
\name{newhall_daymet_subset}
\alias{newhall_daymet_subset}
\title{Load DAYMET Monthly Data at 1 kilometer Resolution}
\usage{
newhall_daymet_subset(
  x,
  start_year = 1991,
  end_year = 2020,
  force = FALSE,
  DAYMET_PATH = tempdir()
)
}
\arguments{
\item{x}{A \emph{SpatVector}, \emph{SpatRaster}, \emph{SpatExtent}, or any other object type suitable to use with \code{terra::ext()}}

\item{start_year}{integer. First year in range to download}

\item{end_year}{integer. Last year in range to download.}

\item{force}{logical. Force download when files exist in \code{DAYMET_PATH}? Default: \code{FALSE}}

\item{DAYMET_PATH}{Default: \code{file.path(newhall_data_dir("cache"), "DAYMET")}}
}
\value{
A \emph{SpatRaster} object
}
\description{
\verb{newhall_daymet_subset():} Used to create a subset of the DAYMET data corresponding to the extent of an input spatial object \code{x}.
}
\references{
Thornton, M.M., R. Shrestha, Y. Wei, P.E. Thornton, S-C. Kao, and B.E. Wilson. 2022. Daymet: Monthly Climate Summaries on a 1-km Grid for North America, Version 4 R1. ORNL DAAC, Oak Ridge, Tennessee, USA. \doi{https://doi.org/10.3334/ORNLDAAC/2131}
}
\seealso{
\code{\link[=newhall_data_dir]{newhall_data_dir()}}
}
