% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallInput.R
\name{newhall_input_csv_batch}
\alias{newhall_input_csv_batch}
\title{Create batch input CSV file for Newhall model}
\usage{
newhall_input_csv_batch(pathname, append = FALSE)
}
\arguments{
\item{pathname}{path to CSV file to create/append}

\item{append}{append to CSV? Default: \code{FALSE}}
}
\value{
a file written to specified \code{pathname} containing a sample station ID, \code{0} for numeric values and \code{""} for character values
}
\description{
Create batch input CSV file for Newhall model
}
