% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallData.R
\name{newhall_issr800_cache}
\alias{newhall_issr800_cache}
\alias{newhall_issr800_extent}
\alias{newhall_issr800_rast}
\title{Load SoilWeb "ISSR-800" at 800 meter Resolution}
\usage{
newhall_issr800_cache(
  ISSR800_PATH = file.path(newhall_data_dir("cache"), "SoilWeb", "800m"),
  template = newhall_prism_rast(),
  overwrite = FALSE
)

newhall_issr800_extent(x)

newhall_issr800_rast(
  ISSR800_PATH = file.path(newhall_data_dir("cache"), "SoilWeb", "800m"),
  tiffile = list.files(ISSR800_PATH, "\\\\.tif$", full.names = TRUE)
)
}
\arguments{
\item{ISSR800_PATH}{Default: \code{file.path(newhall_data_dir("cache"), "SoilWeb", "800m")}}

\item{template}{Template SpatRaster or target CRS specification for reprojection. Default: \code{newhall_prism_rast()}}

\item{overwrite}{Force download of new cache files? Default: \code{FALSE}.}

\item{x}{A \emph{SpatVector}, \emph{SpatRaster}, \emph{SpatExtent}, or any other object type suitable to use with \code{terra::crop()};}

\item{tiffile}{Optional: custom vector of paths to files to use to build raster. Defaults to all .TIF files in the specified cache directory and resolution.}
}
\value{
character. Path to cached water storage GeoTIFF.
}
\description{
Currently the only ISSR-800 data are only available for the contiguous (lower 48) United States. The only property cached for use in the Newhall model is the "available water holding capacity" (sum of storage for the whole profile). For consistency with PRISM grid the values are reprojected from \code{"EPSG:5070"} to \code{"EPSG:4269"}.

\verb{newhall_issr800_extent():} Used to create a subset of the ISSR-800 soil available water storage data corresponding to the extent of an input spatial object \code{x}.

\code{newhall_issr800_rast()}: Create a \emph{SpatRaster} object. This object contains Available Water Capacity (Storage) for at 800 meter resolution using the standard jNSM column naming scheme.
}
\references{
Walkinshaw, Mike, A.T. O'Geen, D.E. Beaudette. "Soil Properties." California Soil Resource Lab, 1 Oct. 2022, \url{https://casoilresource.lawr.ucdavis.edu/soil-properties/}.
}
\seealso{
\code{\link[=newhall_data_dir]{newhall_data_dir()}}
}
