% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallData.R
\name{newhall_prism_cache}
\alias{newhall_prism_cache}
\alias{newhall_prism_rast}
\alias{newhall_prism_subset}
\alias{newhall_nad83_template}
\title{Load PRISM Monthly "Normals" at 800 meter or 4 kilometer Resolution}
\usage{
newhall_prism_cache(
  resolution = "800m",
  overwrite = FALSE,
  PRISM_PATH = file.path(newhall_data_dir("cache"), "PRISM")
)

newhall_prism_rast(
  resolution = "800m",
  PRISM_PATH = file.path(newhall_data_dir("cache"), "PRISM"),
  bilfile = list.files(file.path(PRISM_PATH, resolution), "\\\\.bil$", recursive = TRUE)
)

newhall_prism_subset(
  x,
  resolution = "800m",
  template = newhall_nad83_template(resolution = resolution),
  PRISM_PATH = file.path(newhall_data_dir("cache"), "PRISM")
)

newhall_nad83_template(resolution = "800m")
}
\arguments{
\item{resolution}{character. Either \code{"800m"} (default) or \code{"4km"}}

\item{overwrite}{Force download of new cache files? Default: \code{FALSE}.}

\item{PRISM_PATH}{Default: \code{file.path(newhall_data_dir("cache"), "PRISM")}}

\item{bilfile}{Optional: custom vector of paths to files to use to build raster. Defaults to all .BIL files in the specified cache directory and resolution.}

\item{x}{A \emph{SpatVector}, \emph{SpatRaster}, \emph{SpatExtent}, or any other object type suitable to use with \code{terra::crop()}.}

\item{template}{Template \emph{SpatRaster} or target CRS specification for re-projection. Default: \code{newhall_nad83_template()}}
}
\value{
character. Vector of file paths (to PRISM .BIL files).
}
\description{
\code{newhall_prism_cache()}: Uses the \href{https://cran.r-project.org/package=prism}{prism} package to download and cache data at the specified resolution. At this time only monthly grids for 30 year Normals (1991-2020) are supported.

\code{newhall_prism_rast()}: Create a \emph{SpatRaster} object. This object contains temperature and precipitation data for the specified data set, at the specified resolution, using the standard jNSM column naming scheme.

\verb{newhall_prism_subset():} Used to create a subset of the PRISM data corresponding to the extent of an input spatial object \code{x}.

\code{newhall_nad83_template()}: Empty \code{SpatRaster} corresponding to the lower 48 United States PRISM data/extent at the specified resolution.
}
\examples{

newhall_nad83_template()

newhall_nad83_template("4km")
}
\references{
PRISM Climate Group, Oregon State University, \url{https://prism.oregonstate.edu}, data created 4 Feb 2014, accessed 22 Jul 2023.
}
\seealso{
\code{\link[=newhall_data_dir]{newhall_data_dir()}}
}
