% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASICSimulationModel.R
\name{newhall_simulation}
\alias{newhall_simulation}
\title{Run \emph{BASICSimulationModel} simulation with newhall-1.6.1.jar}
\usage{
newhall_simulation(
  dataset,
  waterHoldingCapacity = 200,
  fc = 2.5,
  fcd = 0.66,
  bsm = BASICSimulationModel(),
  toString = FALSE
)
}
\arguments{
\item{dataset}{a \emph{NewhallDataset} \code{jobjRef}}

\item{waterHoldingCapacity}{Default: \code{200}}

\item{fc}{Default: \code{2.5}}

\item{fcd}{Default: \code{0.66}}

\item{bsm}{\code{jobjRef} for \emph{BASICSimulationModel}; Default: \code{BASICSimulationModel()}}

\item{toString}{logical; return \emph{NewhallResults} (Default: \code{FALSE}), or if \code{TRUE} call \code{result.toString()} to get formatted standard output as \emph{character}?}
}
\value{
\emph{NewhallResults} \code{jobjRef}
}
\description{
Run \emph{BASICSimulationModel} simulation with newhall-1.6.1.jar
}
