% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\name{newhall_version}
\alias{newhall_version}
\title{Get Java Newhall JAR file version}
\usage{
newhall_version()
}
\value{
\emph{character} containing version number of Newhall JAR file
}
\description{
Get Java Newhall JAR file version
}
\details{
This is a wrapper around accessing the public string field \code{NSM_VERSION} stored within the main \emph{Newhall} class of the JAR file.
}
