% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallData.R
\name{newhall_worldclim_cache}
\alias{newhall_worldclim_cache}
\alias{newhall_worldclim_rast}
\alias{newhall_worldclim_subset}
\title{Load WorldClim Monthly Averages}
\usage{
newhall_worldclim_cache(
  resolution = "10m",
  version = "2.1",
  overwrite = FALSE,
  WORLDCLIM_PATH = file.path(newhall_data_dir("cache"), "WorldClim")
)

newhall_worldclim_rast(
  resolution = "10m",
  version = "2.1",
  WORLDCLIM_PATH = file.path(newhall_data_dir("cache"), "WorldClim"),
  tiffile = list.files(file.path(WORLDCLIM_PATH, paste0("wc", version, "_", resolution)),
    pattern = "\\\\.tif$", recursive = TRUE)
)

newhall_worldclim_subset(
  x,
  resolution = "10m",
  template = "EPSG:4326",
  WORLDCLIM_PATH = file.path(newhall_data_dir("cache"), "WorldClim")
)
}
\arguments{
\item{resolution}{character. Either \code{"10m"}, \code{"5m"}, \code{"2.5m"} or \code{"30s"}. In minutes (or seconds) of degrees (\code{"EPSG:4326"}).}

\item{version}{character. Version number. Default: \code{"2.1"}. See \code{geodata::worldclim_global()} for details.}

\item{overwrite}{Force download of new cache files? Default: \code{FALSE}.}

\item{WORLDCLIM_PATH}{Default: \code{file.path(newhall_data_dir("cache"), "WorldClim")}}

\item{tiffile}{Optional: custom vector of paths to files to use to build raster. Defaults to all .TIF files in the specified cache directory and resolution.}

\item{x}{A \emph{SpatVector}, \emph{SpatRaster}, \emph{SpatExtent}, or any other object type suitable to use with \code{terra::crop()}.}

\item{template}{Template \emph{SpatRaster} or target CRS specification for re-projection. Default: \verb{"EPSG:4326}}
}
\value{
character. Vector of file paths (to WorldClim .TIF files).
}
\description{
\code{newhall_worldclim_cache()}: Uses the \href{https://cran.r-project.org/package=geodata}{geodata} package to download and cache data at the specified resolution.

\code{newhall_worldclim_rast()}: Create a \emph{SpatRaster} object. This object contains temperature and precipitation data for the specified data set, at the specified resolution, using the standard jNSM column naming scheme.

\verb{newhall_worldclim_subset():} Used to create a subset of the WorldClim data corresponding to the extent of an input spatial object \code{x}.
}
\references{
Fick, S.E. and R.J. Hijmans, 2017. WorldClim 2: new 1km spatial resolution climate surfaces for global land areas. International Journal of Climatology 37 (12): 4302-4315. \url{https://www.worldclim.org/data/worldclim21.html}
}
\seealso{
\code{\link[=newhall_data_dir]{newhall_data_dir()}}
}
