% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{newhall_XML_export}
\alias{newhall_XML_export}
\title{Export Newhall Results, Data and Metadata to XML file}
\usage{
newhall_XML_export(pathname, dataset, results)
}
\arguments{
\item{pathname}{output XML file path}

\item{dataset}{\emph{NewhallDataset} \code{jobjRef}}

\item{results}{\emph{NewhallResults} \code{jobjRef}}
}
\value{
an XML file written to specified path
}
\description{
Export Newhall Results, Data and Metadata to XML file
}
