% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{newhall_XML_string_export}
\alias{newhall_XML_string_export}
\title{Export Newhall Results, Data and Metadata to XML string}
\usage{
newhall_XML_string_export(dataset, results)
}
\arguments{
\item{dataset}{\emph{NewhallDataset} \code{jobjRef}}

\item{results}{\emph{NewhallResults} \code{jobjRef}}

\item{pathname}{output XML file path}
}
\value{
an XML file written to specified path
}
\description{
Export Newhall Results, Data and Metadata to XML string
}
