% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{newhall_XMLStringResultsExporter}
\alias{newhall_XMLStringResultsExporter}
\title{Export Newhall Results, Data and Metadata to XML string with \emph{XMLStringResultsExporter}}
\usage{
newhall_XMLStringResultsExporter(dataset, results)
}
\arguments{
\item{dataset}{\emph{NewhallDataset} \code{jobjRef}}

\item{results}{\emph{NewhallResults} \code{jobjRef}}
}
\value{
\emph{character} containing XML string
}
\description{
Export Newhall Results, Data and Metadata to XML string with \emph{XMLStringResultsExporter}
}
