% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{NewhallDataset}
\alias{NewhallDataset}
\title{Create an instance of NewhallDataset}
\usage{
NewhallDataset(
  stationName,
  country,
  lat,
  lon,
  nsHemisphere,
  ewHemisphere,
  stationElevation,
  allPrecipsDbl,
  allAirTempsDbl,
  pdbegin,
  pdend,
  smcsawc
)
}
\arguments{
\item{stationName}{character; station name}

\item{country}{character; country}

\item{lat}{double; latitude}

\item{lon}{double; longitude}

\item{nsHemisphere}{character (length 1);\code{'N'} OR \code{'S'}}

\item{ewHemisphere}{character (length 1); \code{'E'} OR \code{'W'}}

\item{stationElevation}{double; station elevation}

\item{allPrecipsDbl}{\verb{java.util.List<Double>} precipitation, monthly}

\item{allAirTempsDbl}{\verb{java.util.List<Double>} precipitation, monthly}

\item{pdbegin}{beginning year}

\item{pdend}{ending year}

\item{smcsawc}{soil moisture control section available water capacity}
}
\description{
Create an instance of NewhallDataset
}
