% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{NewhallDatasetFromPath}
\alias{NewhallDatasetFromPath}
\alias{xml_NewhallDataset}
\alias{csv_NewhallDataset}
\title{Create an instance of \emph{NewhallDataset} from XML or CSV file}
\usage{
NewhallDatasetFromPath(pathname, .parser = XMLFileParser)

xml_NewhallDataset(pathname)

csv_NewhallDataset(pathname)
}
\arguments{
\item{pathname}{\emph{character} containing \code{pathname}}

\item{.parser}{either \code{XMLFileParser} or \code{CSVFileParser}}
}
\description{
Create an instance of \emph{NewhallDataset} from XML or CSV file
}
