% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openJAR.R
\name{openJAR}
\alias{openJAR}
\title{Open the Java Newhall Graphical User Interface}
\usage{
openJAR(intern = FALSE, wait = FALSE)
}
\arguments{
\item{intern}{passed to \code{system()}; when \code{TRUE} return a character result containing output. Default \code{FALSE} returns invisibly.}

\item{wait}{passed to \code{system()} default \code{FALSE} with \code{intern=FALSE} will not interrupt R session.}
}
\value{
If \code{intern=TRUE} (default), the output of the command, one line per character string. \code{0} if successful. If the command could not be run for any reason, the value is 127 and a warning is issued.
}
\description{
Open the Java Newhall Graphical User Interface
}
\details{
See documentation for \code{system()} return resut for limitations on line length, error conditions, etc.
}
\examples{
\donttest{
# openJAR()
}
}
