% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASICSimulationModel.R
\name{run_simulation}
\alias{run_simulation}
\title{Test using BASICSimulationModel class from newhall-1.6.1.jar}
\usage{
run_simulation(
  dataset,
  waterHoldingCapacity = 200,
  fc = 2.5,
  fcd = 0.66,
  bsm = BASICSimulationModel(),
  toString = FALSE
)
}
\arguments{
\item{dataset}{a \emph{NewhallDataset} \code{jobjRef}}

\item{bsm}{a \emph{BASICSimulationModel} \code{jobjRef}; default: \code{BASICSimulationModel()}}
}
\value{
\emph{NewhallResult} \code{jobjRef}
}
\description{
Test using BASICSimulationModel class from newhall-1.6.1.jar
}
