% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRasterLayers.R
\name{writeRasterLayers}
\alias{writeRasterLayers}
\title{Iterate over multiband output and write as single-layer files}
\usage{
writeRasterLayers(x, output_dir = NULL, ...)
}
\arguments{
\item{x}{character. Path to raster file(s) to split by layer.}

\item{output_dir}{character. Default: \code{NULL} creates new directories in current working directory. Alternately specify a different path for output folders to be created.}

\item{...}{Additional arguments to \code{terra::writeRaster()}.}
}
\value{
New directories are created in \code{output_dir} (or current working directory) based on each input file \code{x}.
}
\description{
This function is a simple wrapper around \code{terra:writeRaster()} that makes it easier
to separate the individual layers of an input or output grid as separate files.
}
\examples{
x <- writeRaster(rast(list(a = rast(matrix(1)), 
                           b = rast(matrix(2)))), "test.tif")
writeRasterLayers("test.tif", "test")
unlink(c("test.tif", "test"), recursive=TRUE)
}
