% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{XMLFileParser}
\alias{XMLFileParser}
\title{Create an instance of XMLFileParser}
\usage{
XMLFileParser(pathname)
}
\arguments{
\item{pathname}{a character containing pathname}
}
\value{
an instance of XMLFileParser class
}
\description{
Create an instance of XMLFileParser
}
