% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{XMLResultsExporter}
\alias{XMLResultsExporter}
\title{Create an instance of XMLResultsExporter}
\usage{
XMLResultsExporter(pathname)
}
\arguments{
\item{pathname}{a character containing pathname}
}
\value{
an instance of XMLResultsExporter class
}
\description{
Create an instance of XMLResultsExporter
}
