% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{CSVFileParser}
\alias{CSVFileParser}
\title{Create an instance of CSVFileParser}
\usage{
CSVFileParser(pathname)
}
\arguments{
\item{pathname}{a character containing pathname}
}
\value{
an instance of CSVFileParser class
}
\description{
Create an instance of CSVFileParser
}
