% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallBatch.R
\name{newhall_batch.default}
\alias{newhall_batch.default}
\alias{newhall_batch}
\alias{newhall_batch.character}
\alias{newhall_batch.SpatRaster}
\alias{newhall_batch.RasterBrick}
\title{Run a Newhall simulation for each row in a CSV file}
\usage{
\method{newhall_batch}{default}(
  .data = NULL,
  unitSystem = "metric",
  soilAirOffset = ifelse(unitSystem \%in\% c("in", "english"), 4.5, 2.5),
  amplitude = 0.66,
  verbose = TRUE,
  toString = TRUE,
  checkargs = TRUE,
  cores = NULL,
  file = NULL,
  nrows = NULL,
  overwrite = NULL
)

\method{newhall_batch}{character}(
  .data,
  unitSystem = "metric",
  soilAirOffset = ifelse(unitSystem \%in\% c("in", "english"), 4.5, 2.5),
  amplitude = 0.66,
  verbose = TRUE,
  toString = TRUE,
  checkargs = TRUE,
  cores = 1,
  file = paste0(tempfile(), ".grd"),
  nrows = nrow(.data),
  overwrite = TRUE
)

\method{newhall_batch}{SpatRaster}(
  .data,
  unitSystem = "metric",
  soilAirOffset = ifelse(unitSystem \%in\% c("in", "english"), 4.5, 2.5),
  amplitude = 0.66,
  verbose = TRUE,
  toString = FALSE,
  checkargs = TRUE,
  cores = 1,
  file = paste0(tempfile(), ".tif"),
  nrows = nrow(.data),
  overwrite = TRUE
)

\method{newhall_batch}{RasterBrick}(
  .data,
  unitSystem = "metric",
  soilAirOffset = ifelse(unitSystem \%in\% c("in", "english"), 4.5, 2.5),
  amplitude = 0.66,
  verbose = TRUE,
  toString = TRUE,
  checkargs = TRUE,
  cores = 1,
  file = paste0(tempfile(), ".grd"),
  nrows = nrow(.data),
  overwrite = TRUE
)
}
\arguments{
\item{.data}{a \emph{data.frame} or \emph{character} vector of paths to CSV files; or a SpatRaster or RasterStack containin the same data elements and names as included in the batch \code{data.frame}/CSV format}

\item{unitSystem}{Default: \code{"metric"} OR \code{"mm"} OR \code{"cm"} use \emph{millimeters} of rainfall (default for the BASIC model); set to \code{unitSystem="english"} OR \code{unitSystem="in"} to transform English (inches of precipitation; degrees Fahrenheit) inputs to metric (millimeters of precipitation; degrees Celsius) before running simulation}

\item{soilAirOffset}{air-soil temperature offset. Conventionally for jNSM: \code{2.5} for metric units (default); \code{4.5} for english units.}

\item{amplitude}{difference in amplitude between soil and air temperature sine waves. Default \code{0.66}}

\item{verbose}{print message about number of simulations and elapsed time}

\item{toString}{call \code{toString()} method on each \emph{NewhallResults} object and store in \code{output} column of result?}

\item{checkargs}{\emph{logical}; check argument length and data types for each run? Default: \code{TRUE}}

\item{cores}{number of cores; used only for processing \emph{SpatRaster} or \emph{Raster*} input}

\item{file}{path to write incremental raster processing output for large inputs that do not fit in memory; passed to \code{terra::writeStart()} and used only for processing \emph{SpatRaster} or \emph{Raster*} input; defaults to a temporary file created by \code{tempfile()} if needed}

\item{nrows}{number of rows to use per block; passed to \code{terra::readValues()} \code{terra::writeValues()}; used only for processing \emph{SpatRaster} or \emph{Raster*} input; defaults to number of rows in dataset if needed}

\item{overwrite}{logical; overwrite \code{file}? passed to \code{terra::writeStart()}; defaults to \code{TRUE} if needed}
}
\value{
When input is a data.frame or character vector of paths to CSV files, result is a a \emph{data.frame} with key model outputs (see details) containing list columns with Java Objects for \emph{NewhallDataset}, \emph{NewhallResults}.  If \code{toString=TRUE} the column \code{output} is a \emph{character} containing the \code{toString()} output from \emph{NewhallResults}

For \code{SpatRaster} input returns a \code{SpatRaster} containing numeric and categorical model outputs. \code{RasterStack} inputs are first converted to \code{SpatRaster}, and a \code{SpatRaster} is returned
}
\description{
Run a Newhall simulation for each row in a CSV file
}
\details{
\subsection{Dry v.s. Moist}{

The concept of "dry" versus "moist" is expressed only semi-quantitatively in the Newhall model with three different categories of moisture being recognized: "moist", "moist/dry" and "dry."

Of interest to the classification of climate regimes of a soil are not only when/where the soil is dry but how that moisture or lack thereof corresponds with prevailing temperature conditions.
}

\subsection{Standard model output fields and their definitions in the latest available JAR file include:}{
\itemize{
\item \code{"annualRainfall"} - sum of monthly precipitation values over the year
\item \code{"waterHoldingCapacity"} - total water storage of soil profile in units of length (\code{mm}). Default: \code{200} millimeters (8 inches) of water storage. This is approximately the average water storage when calculated using SSURGO available water capacities and depths for the soils in CONUS.
\item \code{"annualWaterBalance"} - sum of difference of precipitation and estimated mean potential evapotranspiration by month
\item \code{"annualPotentialEvapotranspiration"} - sum of mean monthly potential evapotranspiration
\item \code{"summerWaterBalance"} - sum of (summer months only) difference of precipitation and estimated mean potential evapotranspiration by month
\item \code{"dryDaysAfterSummerSolstice"} - number of days "dry" after June 21; used in definition of Xeric moisture regime
\item \code{"moistDaysAfterWinterSolstice"} - number of days "dry" after December 21; used in definition of Xeric moisture regime
\item \code{"numCumulativeDaysDry"} - cumulative number of "dry" days per year
\item \code{"numCumulativeDaysMoistDry"} - cumulative number of days "intermediate between moist and dry" per year
\item \code{"numCumulativeDaysMoist"} - cumulative number of days "moist" per year
\item \code{"numCumulativeDaysDryOver5C"} - cumulative number of days dry per year when the soil temperature is over 5 degrees C
\item \code{"numCumulativeDaysMoistDryOver5C"} - cumulative number of days intermediate between moist and dry per year when the soil temperature is over 5 degrees C
\item \code{"numCumulativeDaysMoistOver5C"} - cumulative number of days moist per year when the soil temperature is over 5 degrees C
\item \code{"numConsecutiveDaysMoistInSomeParts"} - maximum number of consecutive days per year where some parts of the profile are moist
\item \code{"numConsecutiveDaysMoistInSomePartsOver8C"} - maximum number of consecutive days per year where some parts of the profile are moist and the soil temperature is over 8 degrees C
\item \code{"temperatureRegime"} - estimated Soil Temperature Regime; one of "Pergelic", "Cryic", "Frigid", "Mesic", "Thermic", "Hyperthermic", "Isofrigid", "Isomesic", "Isothermic", or "Isohyperthermic"
\item \code{"moistureRegime"} - estimated Soil Moisture Regime; one of "Aridic", "Ustic", "Xeric", "Udic", "Perudic", or "Undefined"
\item \code{"regimeSubdivision1"} - estimated "Moisture Regime Subdivision #1"; one of "Typic", "Weak", "Wet", "Dry", "Extreme", "Xeric", "Udic", "Aridic", or " " (See van Wambecke and Newhall, 1981)
\item \code{"regimeSubdivision2"} - estimated "Moisture Regime subdivision #2"; one of "Aridic", "Tempustic", "Tropustic", "Tempudic", "Xeric", "Udic", "Tropudic", "Undefined",  or " " (See van Wambecke and Newhall, 1981)
}

"Years" are based on uniform 12 months with 30 days each for a total of 360 days (no leap years).

The following elements have a many:1 relationship with model runs and are not (yet) included in the standard output, but can be accessed using an rJava object reference to a \code{NewhallResults} class.
\itemize{
\item \code{"meanPotentialEvapotranspiration"} - estimated mean monthly potential evapotranspiration
\item \code{"temperatureCalendar"} - compressed (360 day) grid "calendar" showing days above 5 and 8 degrees C
\item \code{"moistureCalendar"} - compressed (360 day) grid "calendar" showing moist, moist/dry and dry days.
}
}
}
\references{
van Wambeke, A. and Newhall, F. and United States Soil Management Support Services (1981) Calculated Soil Moisture and Temperature Regimes of South America: A Compilation of Soil Climatic Regimes calculated by using a mathematical model developed by F. Newhall (Soil Conservation Service, USDA, 1972). SMSS : Technical Monograph : Soil management support services. New York State College of Agriculture and Life Sciences, Cornell University, Department of Agronomy. Available online: \url{https://books.google.com/books?id=jwtIAAAAYAAJ}
}
\seealso{
csv_writeNewhallBatch
}
