% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallGUI.R
\name{newhall_GUI}
\alias{newhall_GUI}
\title{Open the Java Newhall Graphical User Interface}
\usage{
newhall_GUI(command_only = FALSE)
}
\arguments{
\item{command_only}{If \code{TRUE} return the command that would be executed to run GUI}
}
\value{
If \code{intern=TRUE} (default), the output of the command, one line per character string. \code{0} if successful. If the command could not be run for any reason, the value is 127 and a warning is issued.
}
\description{
This function must be called interactively.
}
\details{
See documentation for \code{system()} return result for limitations on line length, error conditions, etc.
}
