% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASICSimulationModel.R
\name{newhall_simulation}
\alias{newhall_simulation}
\title{Run Newhall \emph{BASICSimulationModel} simulation}
\usage{
newhall_simulation(
  dataset,
  smcsawc = 200,
  soilAirOffset = 2.5,
  amplitude = 0.66,
  bsm = BASICSimulationModel(),
  toString = FALSE
)
}
\arguments{
\item{dataset}{a \emph{NewhallDataset} \code{jobjRef}}

\item{smcsawc}{Default: \code{200}}

\item{soilAirOffset}{air-soil temperature offset. Conventionally for jNSM: \code{2.5} for metric units (default); \code{4.5} for english units.}

\item{amplitude}{Default: \code{0.66} amplitude difference between soil and air temperature sine waves}

\item{bsm}{\code{jobjRef} for \emph{BASICSimulationModel}; Default: \code{BASICSimulationModel()}}

\item{toString}{logical; return \emph{NewhallResults} (Default: \code{FALSE}), or if \code{TRUE} call \code{result.toString()} to get formatted standard output as \emph{character}?}
}
\value{
\emph{NewhallResults} \code{jobjRef}
}
\description{
Run Newhall \emph{BASICSimulationModel} simulation
}
