% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{newhall_XMLResultsExporter}
\alias{newhall_XMLResultsExporter}
\title{Export Newhall Input and Output with XMLResultsExporter}
\usage{
newhall_XMLResultsExporter(dataset, results, pathname)
}
\arguments{
\item{dataset}{\emph{NewhallDataset} \code{jobjRef}}

\item{results}{\emph{NewhallResults} \code{jobjRef}}

\item{pathname}{output XML file path}
}
\value{
an XML file written to pathname
}
\description{
Export Newhall Input and Output with XMLResultsExporter
}
