% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{NewhallDataset}
\alias{NewhallDataset}
\title{Create an instance of \emph{NewhallDataset}}
\usage{
NewhallDataset(
  stationName,
  country,
  lat,
  lon,
  nsHemisphere,
  ewHemisphere,
  stationElevation,
  allPrecipsDbl,
  allAirTempsDbl,
  pdbegin,
  pdend,
  smcsawc
)
}
\arguments{
\item{stationName}{\emph{character}; station name}

\item{country}{\emph{character}; country}

\item{lat}{\emph{double}; latitude decimal degrees}

\item{lon}{\emph{double}; longitude decimal degrees}

\item{nsHemisphere}{\emph{character} (length \code{1});\code{'N'} OR \code{'S'}}

\item{ewHemisphere}{\emph{character} (length \code{1}); \code{'E'} OR \code{'W'}}

\item{stationElevation}{\emph{double}; station elevation}

\item{allPrecipsDbl}{\emph{double}; length \code{12} precipitation, monthly}

\item{allAirTempsDbl}{\emph{double}; length \code{12} air temperature, monthly}

\item{pdbegin}{\emph{integer}; beginning year}

\item{pdend}{\emph{integer}; ending year}

\item{smcsawc}{\emph{double}; soil moisture control section available water capacity}
}
\value{
an instance of \emph{NewhallDataset}
}
\description{
Create an instance of \emph{NewhallDataset}
}
\examples{
input_direct <- NewhallDataset(
  stationName = "WILLIAMSPORT",
  country = "US",
  lat = 41.24,
  lon = -76.92,
  nsHemisphere = 'N',
  ewHemisphere = 'W',
  stationElevation = 158.0,
  allPrecipsDbl = c(
    44.2,
    40.39,
    113.54,
    96.77,
    95.0,
    98.55,
    66.04,
    13.46,
    54.86,
    6.35,
    17.53,
    56.39
  ),
  allAirTempsDbl = c(
    -2.17,
    0.89,
    3.72,
    9.11,
    16.28,
    21.11,
    22.83,
    21.94,
    19.78,
    10.5,
    5.33,
    -1.06
  ),
  pdbegin = 1930,
  pdend = 1930,
  smcsawc = 200.0
)

}
