% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{NewhallDataset}
\alias{NewhallDataset}
\title{Create an instance of NewhallDataset}
\usage{
NewhallDataset(
  stationName,
  country,
  lat,
  lon,
  nsHemisphere,
  ewHemisphere,
  stationElevation,
  allPrecipsDbl,
  allAirTempsDbl,
  pdbegin,
  pdend,
  smcsawc
)
}
\arguments{
\item{stationName}{character; station name}

\item{country}{character; country}

\item{lat}{double; latitude}

\item{lon}{double; longitude}

\item{nsHemisphere}{character (length 1);\code{'N'} OR \code{'S'}}

\item{ewHemisphere}{character (length 1); \code{'E'} OR \code{'W'}}

\item{stationElevation}{double; station elevation}

\item{allPrecipsDbl}{precipitation, monthly}

\item{allAirTempsDbl}{precipitation, monthly}

\item{pdbegin}{beginning year}

\item{pdend}{ending year}

\item{smcsawc}{soil moisture control section available water capacity}
}
\description{
Create an instance of NewhallDataset
}
\examples{
input_direct <- NewhallDataset(
  stationName = "WILLIAMSPORT",
  country = "US",
  lat = 41.24,
  lon = -76.92,
  nsHemisphere = 'N',
  ewHemisphere = 'W',
  stationElevation = 158.0,
  allPrecipsDbl = c(
    44.2,
    40.39,
    113.54,
    96.77,
    95.0,
    98.55,
    66.04,
    13.46,
    54.86,
    6.35,
    17.53,
    56.39
  ),
  allAirTempsDbl = c(
    -2.17,
    0.89,
    3.72,
    9.11,
    16.28,
    21.11,
    22.83,
    21.94,
    19.78,
    10.5,
    5.33,
    -1.06
  ),
  pdbegin = 1930,
  pdend = 1930,
  smcsawc = 200.0
)

}
