% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{NewhallDatasetFromPath}
\alias{NewhallDatasetFromPath}
\alias{xml_NewhallDataset}
\alias{csv_NewhallDataset}
\title{Create an instance of NewhallDataset from XML file}
\usage{
NewhallDatasetFromPath(pathname, .parser = XMLFileParser)
}
\arguments{
\item{pathname}{a character containing pathname}

\item{.parser}{either \code{XMLFileParser} or \code{CSVFileParser}}
}
\description{
Create an instance of NewhallDataset from XML file
}
