% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDatasetMetadata.R
\name{NewhallDatasetMetadata}
\alias{NewhallDatasetMetadata}
\title{Create an instance of NewhallDatasetMetadata}
\usage{
NewhallDatasetMetadata(
  stationName,
  stationId,
  stationElevation,
  stationStateProvidence,
  stationCountry,
  mlraName,
  mlraId,
  contribFirstName,
  contribLastName,
  contribTitle,
  contribOrg,
  contribAddress,
  contribCity,
  contribStateProvidence,
  contribPostal,
  contribCountry,
  contribEmail,
  contribPhone,
  notes,
  runDate,
  modelVersion,
  unitSystem,
  soilAirOffset,
  amplitude,
  network
)
}
\arguments{
\item{stationName}{character; station name}

\item{stationId}{character; station ID}

\item{stationElevation}{double; station elevation}

\item{stationStateProvidence}{character; station state / providence}

\item{stationCountry}{character; station country}

\item{mlraName}{character; Major Land Resource Area (MLRA) name}

\item{mlraId}{integer; Major Land Resource Area ID}

\item{contribFirstName}{character; contributor first name}

\item{contribLastName}{character; contributor last name}

\item{contribTitle}{character; contributor title}

\item{contribOrg}{character; contributor organization}

\item{contribAddress}{character; contributor address}

\item{contribCity}{character; contributor city}

\item{contribStateProvidence}{character; contributor state / providence}

\item{contribPostal}{character; contributor postal code}

\item{contribCountry}{character; contributor country}

\item{contribEmail}{character; contributor email}

\item{contribPhone}{character; contributor phone}

\item{notes}{character (may have length >1); notes}

\item{runDate}{character; run date}

\item{modelVersion}{character; model version}

\item{unitSystem}{character; unit system either "cm" or "in"}

\item{soilAirOffset}{double; soil-air temperature offset}

\item{amplitude}{double; soil-air temperature amplitude}

\item{network}{character; network}
}
\value{
an instance of NewhallDatasetMetadata
}
\description{
Create an instance of NewhallDatasetMetadata
}
