% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallDataset.R
\name{XMLFileParser}
\alias{XMLFileParser}
\title{Create an instance of \emph{XMLFileParser}}
\usage{
XMLFileParser(pathname)
}
\arguments{
\item{pathname}{\emph{character} containing \code{pathname}}
}
\value{
an instance of \emph{XMLFileParser} class
}
\description{
Create an instance of \emph{XMLFileParser}
}
