% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{XMLResultsExporter}
\alias{XMLResultsExporter}
\title{Create an instance of \emph{XMLResultsExporter}}
\usage{
XMLResultsExporter(pathname)
}
\arguments{
\item{pathname}{character; output path}
}
\value{
an instance of \emph{XMLResultsExporter} class
}
\description{
Create an instance of \emph{XMLResultsExporter}
}
