% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{XMLStringResultsExporter}
\alias{XMLStringResultsExporter}
\title{Create an instance of \emph{XMLStringResultsExporter}}
\usage{
XMLStringResultsExporter()
}
\value{
an instance of \emph{XMLStringResultsExporter} class
}
\description{
Create an instance of \emph{XMLStringResultsExporter}
}
