% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewhallResults.R
\name{XMLStringResultsExporter}
\alias{XMLStringResultsExporter}
\title{Create an instance of XMLStringResultsExporter}
\usage{
XMLStringResultsExporter(pathname)
}
\arguments{
\item{pathname}{a character containing pathname}
}
\value{
an instance of XMLStringResultsExporter class
}
\description{
Create an instance of XMLStringResultsExporter
}
